refactor: clarify Faust generated module boundary
